allow movies to be renamed

+++ movieDB.rb
@@ -65,9 +65,9 @@ post '/addMovie' do
 	redirect '/movies'
 end
 
-post '/editMovie/:title' do
-	movie = Movie.get(params[:title])
-	movie.destroy!
+post '/editMovie/:key' do
+	movie = Movie.get(params[:key])
+	movie.destroy! if !movie.nil?
 	movie = Movie.new(:release_year => params[:release_year].to_i, :length => params[:length].to_i, :mpaa_rating => params[:mpaa_rating], :plot => params[:plot].gsub("\r\n","<br />"))
 	params[:cast].split(/[ ]?,[ ]?/).each do |name|
 		actor = Actor.get(name)
